+++ packages/nocodb/src/lib/meta/api/helpers/index.ts
@@ -1,3 +1,4 @@
 import { populateMeta } from './populateMeta';
+export * from './columnHelpers';
 
 export { populateMeta };
